+++ src/user_actions/targeting_action.rs
@@ -3,10 +3,10 @@ use rltk::{Rltk, VirtualKeyCode};
 pub enum TargetingAction {
   NoAction,
   Exit,
-  Selected(i32)
+  Selected(usize)
 }
 
-pub fn map_input_to_targeting_action(ctx: &mut Rltk, target: Option<i32>) -> TargetingAction {
+pub fn map_input_to_targeting_action(ctx: &mut Rltk, target: Option<usize>) -> TargetingAction {
   if ctx.left_click {
     return match target {
       Some(idx) => TargetingAction::Selected(idx),
